Move TypeScript input out of CLI to quicktype core
